+++ packages/devtools/bin/cli.js
@@ -2,8 +2,8 @@
 
 import { startViewer } from '../dist/viewer/server.js';
 
-const port = process.env.AI_SDK_DEVTOOLS_PORT
-  ? parseInt(process.env.AI_SDK_DEVTOOLS_PORT)
+const port = process.env.AI_TOOLKIT_DEVTOOLS_PORT
+  ? parseInt(process.env.AI_TOOLKIT_DEVTOOLS_PORT)
   : 4983;
 
 startViewer(port);